Optimizer: convert the FunctionUsesDumper pass to a test
