+++ include/swift/SILOptimizer/PassManager/Passes.def
@@ -158,7 +158,7 @@ MODULE_PASS(MandatoryAllocBoxToStack, "mandatory-allocbox-to-stack",
      "Mandatory stack promotion of box objects")
 MODULE_PASS(AsyncDemotion, "async-demotion",
      "Convert async functions to be synchronous")
-MODULE_PASS(FunctionUsesDumper, "dump-function-uses",
+MODULE_PASS(FunctionUsesTestPass, "test-function-uses",
      "Dump the results of FunctionUses")
 MODULE_PASS(MandatoryPerformanceOptimizations, "mandatory-performance-optimizations",
      "Performs optimizations for performance-annotated functions")
